requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
reportlab==4.0.7
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.11.0
qrcode==7.4.2
Pillow==10.1.0
requests==2.31.0
